Clarify pre-battle template defaults
Expose built-in scenario objective, battlefield, force-size, and BV defaults before player edits while extracting the pre-battle rules and template cards out of the sections barrel.

Constraint: Keep the existing PreBattlePage.sections import surface stable through re-exports.
Rejected: Inline card expansion in the sections barrel | would keep the touched module over the max-lines warning.
Confidence: high
Scope-risk: narrow
Directive: Put additional pre-battle controls in focused components instead of growing the sections barrel.
Tested: npm.cmd test -- --watchAll=false --runTestsByPath src/components/gameplay/pages/__tests__/addSkirmishSetupUI.smoke.test.tsx --runInBand; npm.cmd run lint -- src/components/gameplay/pages/PreBattlePage.sections.tsx src/components/gameplay/pages/ScenarioRulesEditor.tsx src/components/gameplay/pages/ScenarioTemplateCard.tsx src/components/gameplay/pages/__tests__/addSkirmishSetupUI.smoke.test.tsx; npm.cmd run format:check -- src/components/gameplay/pages/PreBattlePage.sections.tsx src/components/gameplay/pages/ScenarioRulesEditor.tsx src/components/gameplay/pages/ScenarioTemplateCard.tsx src/components/gameplay/pages/__tests__/addSkirmishSetupUI.smoke.test.tsx; npm.cmd run typecheck -- --pretty false; npm.cmd run qc:wave3:validate; npm.cmd run qc:app-completion:release -- --json; npm.cmd run qc:lifecycle:status -- --json; npm.cmd run build; git diff --cached --check
Not-tested: Browser E2E visual review of the pre-battle card.
